requirements.txt
langgraph
langchain
langchain-community
langchain-core
chromadb
pydantic
pyyaml
openai
sentence-transformers
simpleeval
instructor
jsonpath-ng
ipykernel
notebook
mcp
pytest
pytest-cov
